[audit] Staked ETH Rounding - Disable exact output (#460)
* fix: steth rounding

* feat: add fork tests

* feat: fork tests

* fix(audit): revert exactOutput swaps for wstETH<>stETH

Due to an unavoidable property of the share mechanics used internally within the stETH implementation, it is not currently possible for v4 to support exact outputs in either direction when interacting with stETH. Rather than violate the spirit of the exactOut request by either overshooting the conversion target or ignoring the unexpected delta, this implementation opts to disallow exactOut by reverting in the `beforeSwap` hook. This resolves H-01.

* feat: add wsteth fork tests

since it has a very specific rounding implementation we test it with the
real thing

* Use infura for fork testing

* add infura api key to environment

* bump deprecated ubuntu version

* Only run fork tests if API key is present, else skip (#462)

* fix: comments

* revert: semgrep

* fix: merge

* fix: snapshots

* fix: snapshots

---------

Co-authored-by: Chris Cashwell <chris.cashwell@uniswap.org>
Co-authored-by: gretzke <daniel@gretzke.de>